# backend/contracts/04-api-contract.openapi.yaml
openapi: 3.0.3
info:
  title: ERP/CRM Backend API
  version: 0.1.0
  description: |
    Контракт backend API для первого этапа внедрения NestJS backend-слоя.

    Backend является источником истины для авторизации, прав доступа, транзакционного
    сохранения заказов, финансовых расчётов, экспорта, VLM-интеграций и аудита.

    Frontend не должен напрямую управлять транзакциями заказа, refresh-token'ами,
    финансовыми итогами, экспортными payload'ами и проверкой прав доступа.

    Stage-1 backend API использует явное path versioning. Текущая стабильная
    версия: /api/v1. Неломающие поля можно добавлять в v1; breaking changes
    требуют /api/v2 и отдельного cutover/deprecation плана.
servers:
  - url: http://localhost:3000
    description: Local NestJS backend
  - url: https://api.example.com
    description: Production backend placeholder
...
paths:
...
  /api/v1/orders/{orderId}/production-stage-events/{productionStatusId}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
      - $ref: '#/components/parameters/ProductionStatusId'
    put:
      tags:
        - Production Actions
      operationId: activateProductionStage
      summary: Активировать производственный этап заказа
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_production_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionStageEventRequest'
      responses:
        '200':
          description: Производственный этап активирован
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Production Actions
      operationId: deactivateProductionStage
      summary: Деактивировать производственный этап заказа
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_production_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionStageEventRequest'
      responses:
        '200':
          description: Производственный этап деактивирован
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/OrderVersionConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/order-details/{detailId}/production-stage-events/{productionStatusId}:
    parameters:
      - $ref: '#/components/parameters/OrderDetailId'
      - $ref: '#/components/parameters/ProductionStatusId'
    put:
      tags:
        - Production Actions
      operationId: activateDetailProductionStage
      summary: Активировать производственный этап детали заказа
      description: |
        Команда идемпотентно фиксирует событие производственного этапа для детали заказа.
        Версия заказа не требуется: повтор той же команды не создает повторный audit/outbox.
      security:
        - bearerAuth: []
      x-permissions:
        - orders.change_production_status
        - orders.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailProductionStageEventRequest'
      responses:
        '200':
          description: Производственный этап детали активирован
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/FeatureDisabled'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token из /api/v1/auth/login или /api/v1/auth/refresh.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh token cookie. Не доступен frontend JavaScript-коду.
  headers:
    RequestId:
      description: Уникальный id запроса для логов и аудита.
      schema:
        type: string
        example: req_01HZX8K5T3M2F5F7K9A2S1QW3E
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    OrderDetailId:
      name: detailId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    ProductionStatusId:
      name: productionStatusId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Некорректные path/query/header параметры
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            badRequest:
              value:
                error:
                  code: BAD_REQUEST
                  message: Некорректные входные данные
                  details:
                    header: If-Match
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            authRequired:
              value:
                error:
                  code: AUTH_REQUIRED
                  message: Требуется авторизация
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            permissionDenied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: Недостаточно прав для выполнения действия
                  details:
                    requiredPermission: orders.update
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    OrderNotFound:
      description: Заказ не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            notFound:
              value:
                error:
                  code: ORDER_NOT_FOUND
                  message: Заказ не найден
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    OrderVersionConflict:
      description: Конфликт версии заказа
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            versionConflict:
              value:
                error:
                  code: ORDER_VERSION_CONFLICT
                  message: Заказ был изменён другим пользователем
                  details:
                    currentVersion: 8
                    clientVersion: 7
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    IdempotencyConflict:
      description: Конфликт idempotency key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            idempotencyKeyReused:
              value:
                error:
                  code: IDEMPOTENCY_KEY_REUSED
                  message: Idempotency key was reused with a different request
                  details:
                    idempotencyKey: detail-production-stage:01HZX8K5T3M2F5F7K9A2S1QW3E
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    ValidationError:
      description: Ошибка валидации payload/query params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Некорректные входные данные
                  details:
                    fields:
                      - path: header.clientId
                        message: clientId обязателен
                      - path: details.0.quantity
                        message: quantity должен быть больше 0
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    FeatureDisabled:
      description: API feature выключена конфигурацией runtime
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            featureDisabled:
              value:
                code: SERVICE_UNAVAILABLE
                message: API feature is disabled
                details:
                  feature: deadlines
                requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Внутренняя ошибка сервера
                  requestId: req_01HZX8K5T3M2F5F7K9A2S1QW3E
  schemas:
    ApiErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Некорректные входные данные
        details:
          nullable: true
          description: Дополнительные сведения об ошибке. Не должен содержать secrets.
        requestId:
          type: string
          example: req_01HZX8K5T3M2F5F7K9A2S1QW3E
...
    ProductionStageEventRequest:
      type: object
      required:
        - version
        - idempotencyKey
      properties:
        version:
          type: integer
          minimum: 0
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 200
    DetailProductionStageEventRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 200
        note:
          type: string
          nullable: true
          maxLength: 1000
    ProductionActionResponse:
      type: object
      required:
        - order
        - requestId
      properties:
        order:
          type: object
          required:
            - orderId
            - version
          properties:
            orderId:
              type: integer
              format: int64
            plannedCompletionDate:
              type: string
              format: date
              nullable: true
            orderStatusId:
              type: integer
              format: int64
            paymentStatusId:
              type: integer
              format: int64
            productionStatusId:
              type: integer
              format: int64
            version:
              type: integer
              minimum: 0
        event:
          type: object
          nullable: true
          required:
            - productionStatusId
            - active
          properties:
            productionEventId:
              type: integer
              format: int64
            productionStatusId:
              type: integer
              format: int64
            active:
              type: boolean
        auditId:
          type: string
          nullable: true
        requestId:
          type: string
...